travis

+++ user_stories/support/env.rb
@@ -14,23 +14,29 @@ Capybara.default_selector = :css
 
 include HyaccConstants
 
-Before do |scenario|
-  dump_dir = File.join('tmp', File.dirname(scenario.file), File.basename(scenario.file, '.feature'))
-
-  if ARGV.include?(scenario.file)
-    db = Dir.glob(File.join(dump_dir, '*.dump.gz')).first
-    if db
-      system("rake dad:db:load DUMP_FILE=#{db} --quiet")
+unless ENV['CI'] != 'travis'
+  Before do |scenario|
+    dump_dir = File.join('tmp', File.dirname(scenario.file), File.basename(scenario.file, '.feature'))
+  
+    if ARGV.include?(scenario.file)
+      db = Dir.glob(File.join(dump_dir, '*.dump.gz')).first
+      if db
+        system("rake dad:db:load DUMP_FILE=#{db} --quiet")
+      end
+    else
+      # 直前のDBをダンプしておく
+      if @current_feature != scenario.file
+        system("rm -Rf #{dump_dir}")
+        system("rake dad:db:dump DUMP_DIR=#{dump_dir} --quiet")
+        @current_feature = scenario.file
+      end
     end
-  else
-    # 直前のDBをダンプしておく
-    if @current_feature != scenario.file
-      system("rm -Rf #{dump_dir}")
-      system("rake dad:db:dump DUMP_DIR=#{dump_dir} --quiet")
-      @current_feature = scenario.file
-    end
+  
+    resize_window(1280, 720)
   end
+end
 
+Before do |scenario|
   resize_window(1280, 720)
 end
